Fix `onMeasure()` for SurfaceView
Add DEBUG flag that doesn't start logging thread and executes NOP when logging in C++
