+++ cpp/include/helpers/general.hpp
@@ -8,10 +8,18 @@
 
 #define LOG_TAG __FILE__
 
+// Print only on debug builds.
+#ifdef DEBUG
 #define LOGE(...) __android_log_print(ANDROID_LOG_ERROR, LOG_TAG, __VA_ARGS__)
 #define LOGW(...) __android_log_print(ANDROID_LOG_WARN, LOG_TAG, __VA_ARGS__)
 #define LOGD(...) __android_log_print(ANDROID_LOG_DEBUG, LOG_TAG, __VA_ARGS__)
 #define LOGI(...) __android_log_print(ANDROID_LOG_INFO, LOG_TAG, __VA_ARGS__)
+#else
+#define LOGE(...)
+#define LOGW(...)
+#define LOGD(...)
+#define LOGI(...)
+#endif
 
 namespace rive_android
 {
@@ -28,5 +36,9 @@ namespace rive_android
 
 	std::string jstring2string(JNIEnv *env, jstring jStr);
 
+#ifdef DEBUG
+	// luigi: this redirects stderr to android log (probably want to ifdef this out for release)
+	void logThread();
+#endif
 } // namespace rive_android
 #endif
